Refactor database imports and enhance user authentication; add type definitions for delegates, countries, and committees
