+++ app/utils/auth.ts
@@ -1,4 +1,4 @@
-import {delegates} from "../db/db";
+import {delegates} from "../db/index";
 
 export const authenticate = (id: string, password: string) => {
   return delegates.find((delegate) => delegate.id === id && delegate.password === password);
